Fix P2 code review findings: layout bugs, error handling, resource leaks
Layout calculation (Kramer-yue):
- Remove duplicate outlineCellHorizontalInset subtraction in RelationLayout
- Fix integer division before Math.round in SchemaNodeLayout.fold()
  and PrimitiveLayout.measure()
- Remove redundant columnSets.clear() in RelationLayout.compress()

Error handling (Kramer-gsd):
- Throw UnsupportedOperationException on named fragment spreads
- Guard ObjectNode cast against null/NullNode server responses
- Log BadRequestException instead of silently swallowing
- Remove dead throws IOException from evaluate() signature
- Add slf4j-api dependency to kramer-ql

Resource leaks (Kramer-wuw):
- Wrap FileInputStream in try-with-resources in Utils.getBits()
- Store and close JAX-RS Client in SinglePageApp lifecycle

+++ kramer/src/main/java/com/chiralbehaviors/layout/SchemaNodeLayout.java
@@ -271,8 +271,8 @@ public abstract sealed class SchemaNodeLayout permits PrimitiveLayout, RelationL
         return new Fold(aggregate,
                         (cardSum == 0
                          || data.size() == 0) ? 1
-                                              : Math.round(cardSum
-                                                           / data.size()),
+                                              : (int) Math.round((double) cardSum
+                                                                  / data.size()),
                         this);
     }
